requirements.txt
requests
beautifulsoup4
pymongo
python-dotenv
flask
pandas
geopy
scikit-learn